+++ packages/synapse-interface/components/StateManagedBridge/BridgeTransactionButton.tsx
@@ -143,7 +143,7 @@ export const BridgeTransactionButton = ({
       onClick: () => switchNetwork(fromChainId),
       pendingLabel: 'Switching chains',
     }
-  } else if (!isApproved) {
+  } else if (!isApproved && fromValueBigInt > 0 && bridgeQuote?.destQuery) {
     buttonProperties = {
       onClick: approveTxn,
       label: `Approve ${fromToken?.symbol}`,